Refactor invoice methods into separate modules
